Remove KernelRouter, move junk out of SocketSession
The big change here is removing KernelRouter, and moving stuff out of
SocketSession:

- SocketSession no longer has repository or any other access to
  notebooks or kernels directly
- Instead, it has a NotebookManager, which manages SharedNotebook
  instances.
- A SocketSession can open a NotebookRef from a SharedNotebook, which
  gives it APIs for doing stuff with the notebook. All that stuff can
  then also be broadcast to the other sessions that are also using that
  SharedNotebook.

This also contains a lot of the server-side work necessary to support
concurrent notebook usage. In particular, status updates and results get
broadcast to all the sessions that use that notebook. Foreign edits also
get broadcast, and half the transformation work is done. But the client
side will have to do some transformation work as well (TODO)