Clear results from the server instead of when Run button is pressed
Currently, the client clears the results when the run button is pressed. But, now that the results are broadcast, we need to control clearing of previous results as part of that broadcast. Otherwise, other clients will just accumulate more results.

This commit adds a new `Result` message, `ClearResults`, which the kernel emits just before the cell starts running. When the client gets that result, it clears any existing results and error messages.