Revert some UI changes, use `Out` instead of `Out(X)`
With this change, any un-named cell expression just gets assigned to `Out`. You can refer to `Out` and it will refer to the last `Out` created in a cell above the current cell.

This avoids polluting the symbol table with tons of `resCellX` values, but keeps the ability to use an expression you evaluated but forgot to assign to a variable.