Initial implementation of lazy, updating, and streaming reprs
This commit adds client/server communication for lazy, updating, and streaming reprs. It adds support for basic values and structs, as well as DataFrames (which is probably the primary use case for this mechanism).

Currently, there's no real UI for using this stuff - if you click on `Out` and there's a streaming repr (i.e. a dataframe) then you'll see the next 50 items in the stream get pulled from the server and logged to the console. Obviously there is much more to be done with this, but this is a large step (and I wanted to get this PR up before it gets any more massive).