Update symbol state in UI
When you focus a cell, updates the symbol table to show what the scope looks like there.

It is a little odd because when you run a cell, it's really the *next* cell that gets the scope, but we want to show the symbols updating without you clicking to the next cell (don't we?). So on a cell that hasn't been run, it shows you the inherited scope, and on a cell that has been run it shows you the resulting scope. Open to suggestions as to how to handle this better.