Don't throw on closed output
Currently, if something tries to write to stdout after all the code in a cell has run, we throw an error. But, there might still be background threads that still try to write to that stdout. This probably deserves a better solution, but for now we'll just swallow that future output rather than throw.