Some fixes to the symbol table
I guess I didn't think through the symbol table changes enough - there remained some issues and odd behavior which I hope this will address.

For one, we were only sending the visible results from the *last cell of the notebook* when the notebook is loaded. So if you reloaded the page and clicked through the notebook cells, some results just wouldn't show up (unless you ran the cells again). This is fixed by sending all results from all cells.

Another thing is that it's a bit odd when you click the "run" button on a cell and don't see anything happen on the symbol table. This is because the cell doesn't get "selected" when you click anything but the editor inside. This is fixed by making any click on the cell cause that cell to be "selected" (but not necessarily focused).

Also fixed some issues with symbols being duplicated oddly.