Warn if eta-expansion fails
Sometimes a method can fail to be eta-expanded into a function (which we use to share the method with i.e. python). If this happens, don't throw an exception; emit a warning instead and skip the symbol.

Also, added UI support for warnings (since we had never used them before).

Will still try and solve the cyclic reference problem, but this is a good step.

Also fixes the issue with warning markers being in the wrong position.