Fix Python scope issue
We ran into
https://stackoverflow.com/questions/39647566/why-does-python-3-exec-fail-when-specifying-locals

Looks like when we pass in our own `locals` to exec it won't update the
globals dictionary itself, so references to outer scopes (like from
within a function) didn't work.

This change splits up the code and execs each top-level statement
separately, so that globals is properly updated. This is actually what
IPython seems to do under the hood so we seem to be in good company.