Prefer the `key` field of keyboard events
This is because the `code` field of keyboard events is not as useful:

* It doesn't know about the user's keyboard layout, so it will be wrong for people with different layouts from US QUERTY
* It treats keys positionally, e.g. `NumpadEnter` vs `Enter` or `LeftShift` vs `Shift`

Therefore we should use the `key` field, which is semantically useful. Unfortunately it seems the events we get from Monaco don't have this, and it's hidden behind `browserEvent`. I'm not sure why that is, but at least `browserEvent` is there.