Revert change to spark session classloader
Put the spark session back on the notebook classloader, but in a slightly different way.

If the spark session isn't started with the notebook classloader as the current thread context classloader, then results of a class defined by the notebook (or a dependency) can't be deserialized. Using `asContext` causes other errors, I guess because the classloader then changes again. This way just leaves it as that thread's context classloader.

Should probably spin up a new `Executor` to start the Spark Session. TODO.