Better default output display for streams
Instead of the string repr, by default show the schema expando thingy. Also, improve the schema expando thingy rather than just turning the schema into an object and displaying that.

Note that this won't affect what is actually saved to the notebook - it only works while the notebook is still running. Which I think is probably good.

Also, updates Spark to 2.1.3. I think we'll keep building against 2.1 (since it's the lowest common denominator, and higher versions should still work because the binary compat seems fine for the APIs we're actually using in Polynote) but might as well have the latest version of that for local use. I did this because of a bug preventing me from making map schemas to test this change against.

Also fix a couple bugs I found in various places.