Revert scodec-stream upgrade
Initially upgraded this in an effort to support Scala 2.13 cross-building. But that didn't work out, and the new scodec-stream is incompatible with the fs2 version that http4s uses, and we can't upgrade http4s because they stopped publishing for Scala 2.11 and we have to publish for Scala 2.11 because Spark.

\o/