Add python runner script
This adds a `polynote.py` runner script as an alternative to the bash
script. It's easier to reliably find jep from within python, and the
theory is that this will also obviate the need for the `LD_PRELOAD`
stuff (since python must already be running!) and will be better about
paying attention to virtual environments and such.