Add a websocket key
Adds a UUID that is either configured or generated on server start. This UUID is embedded in the index page (which should be protected by browser's origin checks) and must be provided to the websocket URL in order to start a websocket session.

This should mitigate #552 somewhat – though I don't want to give the impression that it makes Polynote "safe" to run on your laptop/desktop, because I don't think I'd ever be confident that it's true.

There's probably more we can do here (origin checks, etc) and maybe the key should be sent in a message instead of in the URL. We can continue to try to improve on this, since it doesn't hurt to be more secure, but again I don't think people should ever consider Polynote (or any other notebook tool) to be "safe".