Fit initial plot to the modal
If the window is too small to fit a 960x480 plot, then the axes are clipped and you have to shrink the plot before you can do anything. It's also easy to miss the fact that the axes are even there, so you might not even realize that you have to shrink the plot.

Attempts to automatically adjust the initial plot size to fit the available size.