Remove use of sun.misc.Cleaner
The `sun.misc.Cleaner` API has been removed as of JDK 9. Currently we're using it to ensure that streaming handles etc. are freed when the associated ValueRepr is GC'ed. This commit changes that to use `finalize` instead of `Cleaner`, which from what I understand is viewed as bad for some reason, but at least it doesn't break on JDK 9+.