Add identity provider interfaces
Adds interfaces for supporting identity and permission checks. The only implemented identity provider is `HeaderIdentityProvider`, which pulls from a header and checks against configured permission specs based on the contents of the header. This is useful mainly for running Polynote behind a proxy which handles authentication and passes headers to Polynote.

But the interfaces should make it easy enough to build new identity providers as Polynote plug-ins.

Future work: Presence messages and UI, maybe HTTP authentication support or something like that.