Fix hanging StructDataEncoder macro
The StructDataEncoder macro relied on `fieldsOf` from shapeless in order to find all the encodable fields of a class. But, in cases where you had a `var` member of a class, that method was returning the setters for the var as fields, which means that the generated StructDataEncoder was ill-typed.

For some reason, when this happened, the JVM would get stuck in the `fillInStackTrace` native method for that error. Since it's a native method, it can't be interrupted, and the result is that the cell just hangs while it makes a futile attempt to get the `ReprsOf` for some value in the cell.

This fixes the macro at least, by filtering out any `Unit`-typed accessors as well as anything that actually has parameters. It doesn't fix the general problem that any error that occurs in that macro can cause everything to hang. I'll make another PR after this which reworks how `ReprsOf` are inferred (at least for Scala cells) - like, maybe we should just remove all the auto-derived stuff and just do that completely in the kernel (directly with the compiler, not through macros) when it's applicable, in addition to the implicit search.