[#766] Python scoping issue
Python `exec(object[, globals[, locals]])` behaves differently
depending on whether or not you pass in a `locals` dictionary. If you
do, it executes the code as if it were embedded in a class definition,
which can cause strange scoping issues to occur (like #766).

The solution is to pass in a single dictionary, which need to be a
special TrackingNamespace so we can still track modifications to the
local namespace as a result of code execution.