Add rudimentary WAL files
This adds rudimentary WAL files for last-resort recovery if something goes terribly wrong. Every incoming update is written to the WAL independently of being applied to the in-memory reference.

In the future I'll add something for replaying a WAL and something for cleaning up old WALs. But in the meantime it seems prudent to start producing them.