Narrow Dispatcher, fix lots o' UI bugs (#1016)
Completes the changes started in #892, fixing a bunch of bugs and simplifying some of the state management.

The main change here is the significant reduction of the role of the Dispatcher, which had become a bit of a dumping ground for functionality that didn't really fit elsewhere. Wherever possible, calls to the Dispatcher have been replaced by state changes (with corresponding actions replaced by state observations).

One of the purposes of the Dispatcher was to handle UI actions that changed state in a controlled fashion. Going through the Dispatcher got around a problem with the state updates wherein updating nested state would trigger observers of parent state which would then trigger updates in other states that could potentially cause an infinite loop (since reacting to the state change happened in the Dispatcher rather than in the state-update logic).

In this PR though, we have a new mechanism for preventing this infinite loop: a state update has a notion of the "path" that was updated, such that when an update happens to a nested key it bubbles up to parent observers but never back down to observers on other paths. This mechanism enables us to trigger state updates much more freely than before.

Since it's safer to modify state, StateHandlers now have a lens method, similar to view, that creates a new StateHandler for a nested key. Views are a lot cheaper to make, since they are read-only and memoized by their parent StateHandler. Note however that StateHandlers created by lens are not (yet?) memoized.

Additionally, the lifecycle of State Views and Handlers has changed. The optional Disposable parameter has been removed from their constructor, because it was difficult to reason about their lifecycle upon creation, especially as these Handlers and Views got passed around (the memoization of Views also rendered this a bit moot as well). Instead, a mandatory Disposable parameter was added to addObserver, since it is easier to reason about the lifecycle in the more granular case of a single Observer function.

This also gets rid of UIActions (like in #986) which were confusing and not very helpful (and didn't really provide the abstraction they were supposed to).

Finally, a bunch of bugs have been fixed (many caused by issues rectified by the changes described above).

Resolves #974 (tested in FF and Chrome and it seems fine).