Remove `contentEditable` from monaco, switch to Monaco's context menu
Previously, the Monaco context menu didn't have cut/copy/paste. So we added `contentEditable` to the monaco editor to make those show up in the native context menu. But, that trick doesn't work anymore and causes weird stuff to happen (not sure if this was a change in Chrome or Monaco).

But, the Monaco context menu now does have those features, so this just switches back to using their context menu instead of the native one, and removes `contentEditable`.