Set Python recursion limit to default
For some reason, the recursion limit is set to 3000 in Polynote (the
Python default seems to be 1000).

I have no idea why the limit is different in Polynote. The `jep` console
on my laptop has the limit set to 1000, as does my python installation.

Anyways, this is a problem with Python 3.7 apparently, because it seems
that we were running into https://bugs.python.org/issue38593 (which
caused the underlying issue with #1144).

Since that bug is limited to Python 3.7, we just set the recursion limit
to 1000 for Python 3.7.