Make PySpark dependency distribution configurable
Some python modules should not be distributed via PySpark for various
reasons. For example, Boto doesn't work when distributed as a zip file
(which is how the PySpark distribution works), see
https://github.com/boto/boto3/issues/1770 for details.

This PR adds a set of modules excluded by default, as well as
configurability of this feature. It can be turned off entirely, or a
different list of exclusions can be provided.