Change implementation of locals when exec'ing Python
The TrackingNamespace we were using to determine the values added or
changed by a Python cell was shown to have a significant decrease in
performance. `timeit.timeit("math.sqrt(1234.0)", globals=globals())`
was about 10x slower in Polynote than in the Python REPL, which was
traced to passing in the TrackingNamespace as globals() rather than a
pure Python dictionary.

Passing in a Python dictionary as globals() required a different way
to determine added and changed values. In this implementation, the
AST of the user code is parsed and all declarations or assignments are
pulled out. These are cross-referenced with the contents of the
globals() dictionary after the code is executed, with safeguards to
ensure proper scoping is followed.

For the curious, __polynote_run__ has been extensively commented :)